combat system: hero attack damages enemy
